@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^4.29.12",
-    "@uploadthing/react": "5.1.0",
+    "@uploadthing/react": "5.3.0",
     "autoprefixer": "10.4.14",
     "next": "13.4.4",
     "next-auth": "^4.22.1",
@@ -20,7 +20,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tailwindcss": "^3.3.2",
-    "uploadthing": "5.1.0",
+    "uploadthing": "5.3.3",
     "zod": "^3.21.4",
     "zod-validation-error": "^1.3.0"
   },
@@ -29,9 +29,17 @@
     "@types/node": "18.16.0",
     "@types/react": "18.2.8",
     "@types/react-dom": "18.2.4",
-    "@uploadthing/eslint-config": "0.1.0",
+    "@uploadthing/eslint-config": "0.2.0",
     "@uploadthing/tsconfig": "0.1.0",
-    "eslint": "^8.42.0",
-    "typescript": "5.1.3"
+    "eslint": "^8.47.0",
+    "typescript": "^5.1.6"
+  },
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "@uploadthing/eslint-config/base",
+      "@uploadthing/eslint-config/nextjs",
+      "@uploadthing/eslint-config/react"
+    ]
   }
 }
