@@ -1,44 +1,41 @@
 {
   "name": "@example/solidstart",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "solid-start dev --port 9898",
     "start": "solid-start start --port 9898",
     "build": "solid-start build",
     "lint": "eslint \"**/*.{ts,tsx}\""
   },
-  "type": "module",
-  "devDependencies": {
-    "@types/node": "18.16.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "autoprefixer": "10.4.14",
-    "eslint": "^8.42.0",
-    "eslint-plugin-solid": "^0.9.4",
-    "postcss": "8.4.22",
-    "solid-start-node": "^0.2.26",
-    "tailwindcss": "^3.3.2",
-    "typescript": "5.1.3",
-    "vite": "^4.3.9"
-  },
   "dependencies": {
     "@solidjs/meta": "^0.28.5",
     "@solidjs/router": "^0.8.2",
-    "@uploadthing/solid": "5.1.0",
+    "@uploadthing/solid": "5.1.2",
     "attr-accept": "^2.2.2",
     "file-selector": "^0.6.0",
     "solid-js": "^1.7.6",
     "solid-start": "^0.2.26",
     "solidjs-dropzone": "^1.0.0",
     "undici": "5.20.0",
-    "uploadthing": "5.1.0",
+    "uploadthing": "5.3.3",
     "zod": "^3.21.4"
   },
-  "engines": {
-    "node": ">=16"
+  "devDependencies": {
+    "@types/node": "18.16.0",
+    "@uploadthing/eslint-config": "0.2.0",
+    "autoprefixer": "10.4.14",
+    "eslint": "^8.47.0",
+    "eslint-plugin-solid": "^0.9.4",
+    "postcss": "8.4.22",
+    "solid-start-node": "^0.2.26",
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.6",
+    "vite": "^4.3.9"
   },
   "eslintConfig": {
     "extends": [
+      "@uploadthing/eslint-config/base",
       "plugin:solid/typescript"
     ]
   }
