@@ -1,24 +1,20 @@
 {
   "name": "@uploadthing/eslint-config",
-  "private": true,
-  "version": "0.1.0",
-  "main": "index.js",
+  "version": "0.2.0",
   "license": "MIT",
   "dependencies": {
-    "@types/eslint": "^8.40.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "eslint-config-next": "13.4.1",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-config-turbo": "^1.10.1",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-react": "7.32.2"
+    "@next/eslint-plugin-next": "^13.4.13",
+    "@types/eslint": "^8.44.2",
+    "@typescript-eslint/eslint-plugin": "^6.3.0",
+    "@typescript-eslint/parser": "^6.3.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-turbo": "^1.10.12",
+    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-react": "^7.33.1",
+    "eslint-plugin-react-hooks": "^4.6.0"
   },
   "devDependencies": {
-    "eslint": "^8.42.0",
-    "typescript": "5.1.3"
-  },
-  "publishConfig": {
-    "access": "public"
+    "eslint": "^8.47.0",
+    "typescript": "^5.1.6"
   }
 }
