@@ -31,14 +31,20 @@
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
-    "@uploadthing/eslint-config": "0.1.0",
+    "@uploadthing/eslint-config": "0.2.0",
     "@uploadthing/tsconfig": "0.1.0",
     "@uploadthing/tsup-config": "0.1.0",
-    "eslint": "^8.42.0",
+    "eslint": "^8.47.0",
     "tsup": "6.7.0",
-    "typescript": "5.1.3"
+    "typescript": "^5.1.6"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "@uploadthing/eslint-config/base"
+    ]
   }
 }
