@@ -1,6 +1,6 @@
 {
   "name": "@uploadthing/react",
-  "version": "5.1.0",
+  "version": "5.3.0",
   "license": "MIT",
   "exports": {
     "./package.json": "./package.json",
@@ -17,6 +17,9 @@
   "files": [
     "dist"
   ],
+  "publishConfig": {
+    "access": "public"
+  },
   "typesVersions": {
     "*": {
       "*": [
@@ -34,7 +37,8 @@
     "dev": "wait-on ../uploadthing/dist/server.mjs && tsup --watch"
   },
   "dependencies": {
-    "@uploadthing/shared": "^5.0.1"
+    "@uploadthing/shared": "^5.2.0",
+    "tailwind-merge": "^1.13.2"
   },
   "peerDependencies": {
     "react": "^17.0.2 || ^18.0.0",
@@ -45,23 +49,27 @@
     "@types/node": "18.16.0",
     "@types/react": "18.2.8",
     "@types/react-dom": "18.2.4",
-    "@uploadthing/eslint-config": "0.1.0",
+    "@uploadthing/eslint-config": "0.2.0",
     "@uploadthing/tsconfig": "0.1.0",
     "@uploadthing/tsup-config": "0.1.0",
     "autoprefixer": "10.4.14",
-    "eslint": "^8.42.0",
+    "eslint": "^8.47.0",
     "postcss": "8.4.22",
     "react": "18.2.0",
     "react-dropzone": "^14.2.3",
     "tailwindcss": "^3.3.2",
     "tsup": "6.7.0",
-    "typescript": "5.1.3",
-    "uploadthing": "5.1.0",
+    "typescript": "^5.1.6",
+    "uploadthing": "5.3.3",
     "vitest": "^0.30.1",
     "wait-on": "^7.0.1",
     "zod": "^3.21.4"
   },
-  "publishConfig": {
-    "access": "public"
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "@uploadthing/eslint-config/base",
+      "@uploadthing/eslint-config/react"
+    ]
   }
 }
