@@ -1,6 +1,6 @@
 {
   "name": "uploadthing",
-  "version": "5.1.0",
+  "version": "5.3.3",
   "license": "MIT",
   "exports": {
     "./package.json": "./package.json",
@@ -20,11 +20,19 @@
     "./next-legacy": {
       "import": "./dist/next-legacy.mjs",
       "types": "./dist/next-legacy.d.ts"
+    },
+    "./tw": {
+      "types": "./dist/tw.d.ts",
+      "import": "./dist/tw.mjs",
+      "require": "./dist/tw.js"
     }
   },
   "files": [
     "dist"
   ],
+  "publishConfig": {
+    "access": "public"
+  },
   "typesVersions": {
     "*": {
       "client": [
@@ -38,6 +46,9 @@
       ],
       "next-legacy": [
         "dist/next-legacy.d.ts"
+      ],
+      "tw": [
+        "dist/tw.d.ts"
       ]
     }
   },
@@ -52,21 +63,25 @@
   },
   "dependencies": {
     "@uploadthing/mime-types": "^0.2.0",
-    "@uploadthing/shared": "^5.0.1"
+    "@uploadthing/shared": "^5.2.0"
   },
   "devDependencies": {
-    "@uploadthing/eslint-config": "0.1.0",
+    "@uploadthing/eslint-config": "0.2.0",
     "@uploadthing/tsconfig": "0.1.0",
     "@uploadthing/tsup-config": "0.1.0",
-    "eslint": "^8.42.0",
+    "eslint": "^8.47.0",
     "next": "13.4.4",
+    "tailwindcss": "^3.3.2",
     "tsup": "6.7.0",
     "type-fest": "^3.11.1",
-    "typescript": "5.1.3",
+    "typescript": "^5.1.6",
     "vitest": "^0.30.1",
     "zod": "^3.21.4"
   },
-  "publishConfig": {
-    "access": "public"
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "@uploadthing/eslint-config/base"
+    ]
   }
 }
