@@ -6,6 +6,7 @@ import { UploadButton, UploadDropzone } from "~/utils/uploadthing";
 
 export default function Home() {
   const [userInput, setUserInput] = useState("");
+  const [error, setError] = useState("");
 
   return (
     <main className="flex min-h-screen flex-col items-center justify-center  gap-16 p-24">
@@ -17,6 +18,7 @@ export default function Home() {
           value={userInput}
           onChange={(e) => setUserInput(e.currentTarget.value)}
         />
+        {error && <div className="text-red-500">{error}</div>}
       </div>
       <div className="flex flex-col items-center justify-center gap-4">
         <div className="flex gap-4">
@@ -27,8 +29,13 @@ export default function Home() {
               console.log("Files: ", res);
               alert("Upload Completed");
             }}
-            onUploadError={(error: Error) => {
-              alert(`ERROR! ${error.message}`);
+            onUploadError={(error) => {
+              console.log("Error: ", error);
+              if (error.data?.zodError?.fieldErrors.foo) {
+                setError(error.data?.zodError?.fieldErrors.foo[0]);
+              } else {
+                setError("");
+              }
             }}
           />
 
@@ -38,7 +45,7 @@ export default function Home() {
               console.log("Files: ", res);
               alert("Upload Completed");
             }}
-            onUploadError={(error: Error) => {
+            onUploadError={(error) => {
               alert(`ERROR! ${error.message}`);
             }}
           />
