@@ -6,7 +6,7 @@ interface State<T> {
   error?: Error;
 }
 
-type Cache<T> = { [url: string]: T };
+type Cache<T> = Record<string, T>;
 
 // discriminated union type
 type Action<T> =
