@@ -3,7 +3,7 @@ import type { OnDropHandler } from "solidjs-dropzone";
 import { createDropzone } from "solidjs-dropzone";
 
 import type { ExpandedRouteConfig } from "@uploadthing/shared";
-import type { UploadFileType } from "uploadthing/client";
+import type { UploadFileResponse } from "uploadthing/client";
 import {
   classNames,
   generateClientDropzoneAccept,
@@ -47,7 +47,6 @@ const INTERNAL_doFormatting = (config?: ExpandedRouteConfig): string => {
   const key = allowedTypes[0];
   const formattedKey = formattedTypes[0];
 
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
   const { maxFileSize, maxFileCount } = config[key]!;
 
   if (maxFileCount && maxFileCount > 1) {
@@ -68,20 +67,19 @@ export type UploadthingComponentProps<TRouter extends FileRouter> = {
     endpoint: TEndpoint;
 
     onUploadProgress?: (progress: number) => void;
-    onClientUploadComplete?: (
-      res?: Awaited<ReturnType<UploadFileType<TRouter>>>,
-    ) => void;
+    onClientUploadComplete?: (res?: UploadFileResponse[]) => void;
     onUploadError?: (error: Error) => void;
     url?: string;
     multiple?: boolean;
   } & (undefined extends inferEndpointInput<TRouter[TEndpoint]>
-    ? {}
+    ? // eslint-disable-next-line @typescript-eslint/ban-types
+      {}
     : {
         input: inferEndpointInput<TRouter[TEndpoint]>;
       });
 }[keyof TRouter];
 
-const progressHeights: { [key: number]: string } = {
+const progressHeights: Record<number, string> = {
   0: "after:ut-w-0",
   10: "after:ut-w-[10%]",
   20: "after:ut-w-[20%]",
@@ -144,7 +142,6 @@ export function UploadButton<TRouter extends FileRouter>(
       >
         <input
           class="ut-hidden"
-          /*eslint-disable @typescript-eslint/no-non-null-assertion*/
           ref={inputRef!}
           type="file"
           multiple={fileInfo().multiple}
