@@ -1,8 +1,9 @@
 /* eslint-disable @typescript-eslint/no-unsafe-argument */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import { pollForFileData } from "@uploadthing/shared";
+import { pollForFileData, UploadThingError } from "@uploadthing/shared";
 
+import { maybeParseResponseXML } from "./internal/s3-error-parser";
 import type { FileRouter, inferEndpointInput } from "./internal/types";
 
 function fetchWithProgress(
@@ -16,7 +17,7 @@ function fetchWithProgress(
 ) {
   return new Promise<XMLHttpRequest>((res, rej) => {
     const xhr = new XMLHttpRequest();
-    xhr.open(opts.method || "get", url);
+    xhr.open(opts.method ?? "get", url);
     opts.headers &&
       Object.keys(opts.headers).forEach(
         (h) =>
@@ -54,6 +55,33 @@ type UploadFilesOptions<TRouter extends FileRouter> = {
   };
 }[keyof TRouter];
 
+export type UploadFileResponse = {
+  /**
+   * @deprecated
+   * use `name` instead
+   */
+  fileName: string;
+  name: string;
+  /**
+   * @deprecated
+   * use `size` instead
+   */
+  fileSize: number;
+  size: number;
+  /**
+   * @deprecated
+   * use `key` instead
+   */
+  fileKey: string;
+  key: string;
+  /**
+   * @deprecated
+   * use `url` instead
+   */
+  fileUrl: string;
+  url: string;
+};
+
 export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
   opts: UploadFilesOptions<TRouter>,
   config?: {
@@ -73,9 +101,12 @@ export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
         input: opts.input,
       }),
     },
-  ).then((res) => {
+  ).then(async (res) => {
     // check for 200 response
-    if (!res.ok) throw new Error("Failed to get presigned URLs");
+    if (!res.ok) {
+      const error = await UploadThingError.fromResponse(res);
+      throw error;
+    }
 
     // attempt to parse response
     try {
@@ -83,19 +114,34 @@ export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
     } catch (e) {
       // response is not JSON
       console.error(e);
-      throw new Error(`Failed to parse response as JSON. Got: ${res.body}`);
+      throw new UploadThingError({
+        code: "BAD_REQUEST",
+        message: `Failed to parse response as JSON. Got: ${await res.text()}`,
+        cause: e,
+      });
     }
   });
 
-  if (!s3ConnectionRes || !Array.isArray(s3ConnectionRes))
-    throw "No url received. How did you get here?";
+  if (!s3ConnectionRes || !Array.isArray(s3ConnectionRes)) {
+    throw new UploadThingError({
+      code: "BAD_REQUEST",
+      message: "No URL. How did you even get here?",
+      cause: s3ConnectionRes,
+    });
+  }
 
   const fileUploadPromises = s3ConnectionRes.map(async (presigned: any) => {
     const file = opts.files.find((f) => f.name === presigned.name);
 
     if (!file) {
       console.error("No file found for presigned URL", presigned);
-      throw new Error("No file found for presigned URL");
+      throw new UploadThingError({
+        code: "NOT_FOUND",
+        message: "No file found for presigned URL",
+        cause: `Expected file with name ${
+          presigned.name
+        } but got '${opts.files.join(",")}'`,
+      });
     }
     const { url, fields } = presigned.presignedUrl;
     const formData = new FormData();
@@ -136,28 +182,47 @@ export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
         }),
     );
 
-    if (upload.status > 299 || upload.status < 200)
-      throw new Error("Upload failed.");
+    if (upload.status > 299 || upload.status < 200) {
+      // Attempt to parse response as XML
+      const parsed = maybeParseResponseXML(upload.responseText);
+
+      if (parsed?.message) {
+        throw new UploadThingError({
+          code: parsed.code,
+          message: parsed.message,
+        });
+      } else {
+        throw new UploadThingError({
+          code: "UPLOAD_FAILED",
+          message: `Failed to upload file ${file.name} to S3`,
+          cause: upload.responseText,
+        });
+      }
+    }
+
     // Generate a URL for the uploaded image since AWS won't give me one
     const genUrl =
-      "https://uploadthing.com/f/" + encodeURIComponent(fields["key"]);
+      "https://uploadthing.com/f/" + encodeURIComponent(fields.key);
 
     // Poll for file data, this way we know that the client-side onUploadComplete callback will be called after the server-side version
     await pollForFileData(presigned.key);
 
-    return {
+    // TODO: remove `file` prefix in next major version
+    const ret: UploadFileResponse = {
+      fileName: file.name,
+      name: file.name,
+      fileSize: file.size,
+      size: file.size,
       fileKey: presigned.key,
+      key: presigned.key,
       fileUrl: genUrl,
+      url: genUrl,
     };
+    return ret;
   });
 
-  return Promise.all(fileUploadPromises) as Promise<
-    { fileUrl: string; fileKey: string }[]
-  >;
+  return Promise.all(fileUploadPromises);
 };
-
-export type UploadFileType<TRouter extends FileRouter> =
-  typeof DANGEROUS__uploadFiles<TRouter>;
 
 export const genUploader = <
   TRouter extends FileRouter,
@@ -165,7 +230,7 @@ export const genUploader = <
   return DANGEROUS__uploadFiles;
 };
 
-export const classNames = (...classes: Array<string | boolean>) => {
+export const classNames = (...classes: (string | boolean)[]) => {
   return classes.filter(Boolean).join(" ");
 };
 
@@ -173,6 +238,7 @@ export const generateMimeTypes = (fileTypes: string[]) => {
   const accepted = fileTypes.map((type) => {
     if (type === "blob") return "blob";
     if (type === "pdf") return "application/pdf";
+    if (type.includes("/")) return type;
     else return `${type}/*`;
   });
 
