@@ -1,7 +1,12 @@
+import type { Json } from "@uploadthing/shared";
+
+import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
 
 export * from "./internal/types";
 export * as utapi from "./sdk";
 export { createServerHandler } from "./internal/edge";
 
-export const createUploadthing = () => createBuilder<"web">();
+export const createUploadthing = <TErrorShape extends Json>(
+  opts?: CreateBuilderOptions<TErrorShape>,
+) => createBuilder<"web", TErrorShape>(opts);
