@@ -2,13 +2,22 @@ import { defineConfig } from "tsup";
 
 import { config } from "@uploadthing/tsup-config";
 
-export default defineConfig((opts) => ({
-  ...config,
-  entry: [
-    "./src/client.ts",
-    "./src/server.ts",
-    "./src/next.ts",
-    "./src/next-legacy.ts",
-  ],
-  clean: !opts.watch,
-}));
+export default defineConfig((opts) => [
+  {
+    ...config,
+    entry: [
+      "./src/client.ts",
+      "./src/server.ts",
+      "./src/next.ts",
+      "./src/next-legacy.ts",
+      "./src/tw.ts",
+    ],
+    clean: !opts.watch,
+  },
+  {
+    ...config,
+    entry: ["./src/tw.ts"],
+    clean: !opts.watch,
+    format: ["cjs", "esm"],
+  },
+]);
